Start doing a bit of refactoring
