+++ src/main/java/dev/ngspace/hudder/hudderv3/V3ExecuteMethodWriter.java
@@ -6,6 +6,8 @@ import org.objectweb.asm.Type;
 
 import dev.ngspace.hudder.Hudder;
 import dev.ngspace.hudder.api.functionsandconsumers.ArrayElementManager;
+import dev.ngspace.hudder.hudderv3.asm.V3ClassWriter;
+import dev.ngspace.hudder.hudderv3.asm.V3MethodWriter;
 
 public class V3ExecuteMethodWriter extends V3MethodWriter {
