+++ src/main/java/dev/ngspace/hudder/hudderv3/asm/V3ClassWriter.java
@@ -1,4 +1,4 @@
-package dev.ngspace.hudder.hudderv3;
+package dev.ngspace.hudder.hudderv3.asm;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -14,6 +14,9 @@ import dev.ngspace.hudder.api.functionsandconsumers.FunctionAndConsumerAPI.Binda
 import dev.ngspace.hudder.api.functionsandconsumers.FunctionAndConsumerAPI.BindableFunction;
 import dev.ngspace.hudder.api.functionsandconsumers.FunctionAndConsumerAPI.Binder;
 import dev.ngspace.hudder.compilers.HudderV3Compiler;
+import dev.ngspace.hudder.hudderv3.GeneratedCompiler;
+import dev.ngspace.hudder.hudderv3.HudderV3Helper;
+import dev.ngspace.hudder.hudderv3.V3ExecuteMethodWriter;
 
 public class V3ClassWriter implements Binder {
 	public ClassWriter classWriter;
