+++ src/main/java/dev/ngspace/hudder/hudderv3/instructions/variables/constants/BooleanVariableVisitor.java
@@ -2,7 +2,7 @@ package dev.ngspace.hudder.hudderv3.instructions.variables.constants;
 
 import dev.ngspace.hudder.compilers.abstractions.AV3Compiler;
 import dev.ngspace.hudder.exceptions.CompileException;
-import dev.ngspace.hudder.hudderv3.V3MethodWriter;
+import dev.ngspace.hudder.hudderv3.asm.V3MethodWriter;
 import dev.ngspace.hudder.hudderv3.instructions.variables.VariableVisitor;
 
 public class BooleanVariableVisitor extends VariableVisitor {
