+++ src/main/java/dev/ngspace/hudder/hudderv3/instructions/variables/modifiable/DynamicVariableVisitor.java
@@ -3,7 +3,7 @@ package dev.ngspace.hudder.hudderv3.instructions.variables.modifiable;
 import dev.ngspace.hudder.compilers.abstractions.AV3Compiler;
 import dev.ngspace.hudder.compilers.abstractions.AVarTextCompiler;
 import dev.ngspace.hudder.exceptions.CompileException;
-import dev.ngspace.hudder.hudderv3.V3MethodWriter;
+import dev.ngspace.hudder.hudderv3.asm.V3MethodWriter;
 import dev.ngspace.hudder.hudderv3.instructions.variables.VariableVisitor;
 
 public class DynamicVariableVisitor extends VariableVisitor {
