+++ src/main/java/dev/ngspace/hudder/hudderv3/instructions/variables/modifiable/TemporaryVariableVisitor.java
@@ -8,7 +8,7 @@ import org.objectweb.asm.Opcodes;
 import dev.ngspace.hudder.compilers.abstractions.AV3Compiler;
 import dev.ngspace.hudder.compilers.abstractions.AVarTextCompiler;
 import dev.ngspace.hudder.exceptions.CompileException;
-import dev.ngspace.hudder.hudderv3.V3MethodWriter;
+import dev.ngspace.hudder.hudderv3.asm.V3MethodWriter;
 import dev.ngspace.hudder.hudderv3.instructions.variables.VariableVisitor;
 
 public class TemporaryVariableVisitor extends VariableVisitor {
