+++ src/main/java/dev/ngspace/hudder/hudderv3/instructions/variables/operations/ClassAccessVariableVisitor.java
@@ -3,7 +3,7 @@ package dev.ngspace.hudder.hudderv3.instructions.variables.operations;
 import dev.ngspace.hudder.compilers.abstractions.AV3Compiler;
 import dev.ngspace.hudder.exceptions.CompileException;
 import dev.ngspace.hudder.hudderv3.HudderV3Helper;
-import dev.ngspace.hudder.hudderv3.V3MethodWriter;
+import dev.ngspace.hudder.hudderv3.asm.V3MethodWriter;
 import dev.ngspace.hudder.hudderv3.instructions.variables.VariableVisitor;
 import dev.ngspace.hudder.utils.HudderUtils;
