+++ src/main/java/dev/ngspace/hudder/hudderv3/instructions/variables/operations/booloperations/LogicalOrVariableVisitor.java
@@ -5,7 +5,7 @@ import org.objectweb.asm.Opcodes;
 
 import dev.ngspace.hudder.compilers.abstractions.AV3Compiler;
 import dev.ngspace.hudder.exceptions.CompileException;
-import dev.ngspace.hudder.hudderv3.V3MethodWriter;
+import dev.ngspace.hudder.hudderv3.asm.V3MethodWriter;
 import dev.ngspace.hudder.hudderv3.instructions.variables.VariableVisitor;
 
 public class LogicalOrVariableVisitor extends VariableVisitor {
